+++ exonum/src/blockchain/tests.rs
@@ -15,8 +15,8 @@
 use exonum_crypto::{PublicKey, SecretKey};
 use exonum_derive::FromAccess;
 use exonum_merkledb::{
-    access::Access, BinaryValue, Error as MerkledbError, ObjectHash, ProofListIndex, Snapshot,
-    SystemSchema,
+    access::{Access, FromAccess},
+    BinaryValue, Error as MerkledbError, ObjectHash, ProofListIndex, Snapshot, SystemSchema,
 };
 use futures::{Future, IntoFuture};
 use semver::Version;
@@ -71,6 +71,12 @@ fn create_genesis_config() -> GenesisConfig {
 #[derive(Debug, FromAccess)]
 struct InspectorSchema<T: Access> {
     values: ProofListIndex<T::Base, u64>,
+}
+
+impl<T: Access> InspectorSchema<T> {
+    fn new(access: T) -> Self {
+        Self::from_root(access).unwrap()
+    }
 }
 
 /// Actions that performs at the `initiate_adding_service` stage.