+++ exonum/src/proto/mod.rs
@@ -20,8 +20,9 @@ pub use self::schema::{
     consensus::{
         BlockRequest, BlockResponse, Connect, ExonumMessage, PeersRequest, Precommit, Prevote,
         PrevotesRequest, Propose, ProposeRequest, SignedMessage, Status, TransactionsRequest,
-        TransactionsResponse,
+        TransactionsResponse, Verified,
     },
+    proofs::{BlockProof, IndexProof},
     runtime::{AnyTx, CallInfo, GenesisConfig, InstanceInitParams},
 };