+++ services/supervisor/tests/simple.rs
@@ -115,11 +115,8 @@ impl Configure for ConfigChangeService {
 
 fn assert_config_change_is_applied(testkit: &TestKit) {
     let snapshot = testkit.snapshot();
-    assert!(
-        !Schema::new(snapshot.for_service(supervisor_name()).unwrap())
-            .pending_proposal
-            .exists()
-    );
+    let schema: Schema<_> = snapshot.service_schema(supervisor_name()).unwrap();
+    assert!(!schema.pending_proposal.exists());
 }
 
 /// Attempts to change consensus config with only one confirmation.
@@ -321,13 +318,9 @@ fn test_send_proposal_with_api() {
 
     // Assert that config is now pending.
     let snapshot = testkit.snapshot();
-    let snapshot = snapshot.for_service(supervisor_name()).unwrap();
+    let schema: Schema<_> = snapshot.service_schema(supervisor_name()).unwrap();
     assert_eq!(
-        Schema::new(snapshot)
-            .pending_proposal
-            .get()
-            .unwrap()
-            .config_propose,
+        schema.pending_proposal.get().unwrap().config_propose,
         config_propose
     );
 
@@ -366,8 +359,6 @@ fn deploy_service() {
         .query(&deploy_request)
         .post("deploy-artifact")
         .unwrap();
-    testkit.create_block();
-
     testkit.create_blocks_until(deadline_height);
 
     // Check that request was executed.